Cleanup some types and code
We now cleanup some type conversions as well as method signatures.
